# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Post Sync HTTP API
  version: 0.2.0
  summary: Publish one post to Instagram, Facebook, TikTok, YouTube, LinkedIn, Threads, X and Bluesky.
  description: |
    The HTTP API served by `createHandler()` from the `post-social-media-sync` package (and by its Express and Fastify
    adapters), plus the few extra routes of the ready-made Post Sync server (`apps/dashboard`).

    **Paths are relative to where the handler is mounted**: the path of the engine's `publicUrl`. An app that calls
    `createPostSync({ publicUrl: "https://api.example.com/social" })` serves `GET https://api.example.com/social/accounts`.
    The ready-made server mounts it at `<PUBLIC_BASE_URL>/api`.

    **Who is calling.** Every route except the OAuth callback and the signed media files belongs to one *owner*: the
    user or workspace id returned by the `authenticate` hook you pass to the handler. Owners never see each other's
    accounts, media or posts. When you embed the package, authentication is whatever your hook checks (your session
    cookie, a JWT, ...): the security schemes below describe the ready-made server only.
    A missing owner answers `401`; a 4xx error thrown by the hook (such as the ready-made server's `400` for a malformed
    `X-Owner-Id`) is passed through with its status.

    **Conventions**
    - Bodies are JSON (`Content-Type: application/json`, 1 MB max by default, `413` above), except uploads
      (`multipart/form-data`).
...
  license:
...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /platforms:
    parameters:
      - $ref: "#/components/parameters/OwnerId"
    get:
      tags: [Platforms]
      operationId: describe
      summary: Platforms, options, limits and setup state
      description: >-
        Everything a composer UI needs: each platform's capabilities (caption length, media limits) and options, and
        each connector's login kind, whether its app keys are set, and the redirect URI to register.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Description" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ---- accounts -------------------------------------------------------------------------------------------------
...
  /login:
    post:
      tags: [Dashboard server]
      operationId: dashboardLogin
      summary: Log in to the dashboard (DASHBOARD=on only)
      description: >-
        Checks `ADMIN_PASSWORD` and sets the `pss_session` cookie (HttpOnly, SameSite=Lax, 30 days). The session acts
        as the owner `default`. After 5 wrong passwords from one IP, logins from it are refused for 15 minutes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "200":
          description: Logged in.
          headers:
            Set-Cookie:
              description: "`pss_session=<expiry>.<signature>; Path=/; HttpOnly; SameSite=Lax` (+ `Secure` on https)"
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "401":
          description: Wrong password.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: Wrong password. }
        "429":
          description: Too many failed attempts.
          headers:
            Retry-After:
              description: Seconds until logins are accepted again.
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /meta:
    parameters:
      - $ref: "#/components/parameters/OwnerId"
    get:
      tags: [Dashboard server]
      operationId: dashboardMeta
      summary: Description plus setup hints for the dashboard
      description: >-
        Same as `GET /platforms`, plus the server's `PUBLIC_BASE_URL` and, per connector, the environment variables
        that enable it. Available with DASHBOARD on or off.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Description"
                  - type: object
                    required: [publicBaseUrl, connectors]
                    properties:
                      publicBaseUrl:
                        type: string
                        format: uri
                        description: "`PUBLIC_BASE_URL` (the API lives under `<publicBaseUrl>/api`)."
                      connectors:
                        type: array
                        items:
                          allOf:
                            - $ref: "#/components/schemas/ConnectorDescription"
                            - type: object
                              required: [envVars]
                              properties:
                                envVars:
                                  type: array
                                  items: { type: string }
                                  examples: [[META_APP_ID, META_APP_SECRET]]
        "400":
          description: Malformed `X-Owner-Id` header.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Ready-made server: `Authorization: Bearer <API_TOKEN>`. Acts for the owner `default` (the dashboard's data), or
        for any owner named in the `X-Owner-Id` header, so one backend can serve all its users. A wrong token answers
        401 (it doesn't fall back to the cookie). When you embed the package, your `authenticate` hook decides.
    cookieAuth:
      type: apiKey
      in: cookie
      name: pss_session
      description: >-
        Ready-made server with DASHBOARD=on: the session cookie set by `POST /login`. Acts as the owner `default`.
        When you embed the package, use your own session instead.
  parameters:
    OwnerId:
      name: X-Owner-Id
      in: header
      required: false
      description: >-
        Ready-made server with bearer auth only: the owner (your user or workspace id) to act for. Default `default`.
        A value not matching the pattern answers 400. Ignored for cookie sessions. Not part of the package: embedders
        identify the owner in their `authenticate` hook.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._:@|-]{1,200}$"
        examples: [user_42]
...
  responses:
...
    Unauthorized:
      description: Not logged in (the `authenticate` hook returned no owner).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: Not logged in. }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: A message you can show to the user.
        issues:
          type: array
          description: Per-account problems (creating a post).
          items: { $ref: "#/components/schemas/TargetIssue" }
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, const: true }
...
    PlatformId:
      type: string
      enum: [facebook, instagram, threads, tiktok, linkedin, youtube, x, bluesky]
    ConnectorId:
      type: string
      description: >-
        A login that connects one or more platforms: `meta` (Facebook Pages + Instagram), `threads`, `tiktok`,
        `linkedin`, `google` (YouTube), `x`, `bluesky`.
      enum: [meta, threads, tiktok, linkedin, google, x, bluesky]
...
    Capabilities:
      type: object
      required: [textOnly, maxTextLength, maxImages, video, mixedMedia, needsPublicMediaUrl, usesTitle]
      properties:
        textOnly:
          type: boolean
          description: Whether a post with no media is allowed.
        maxTextLength:
          type: integer
          description: Max caption length (as the platform counts characters).
        maxImages:
          type: integer
          description: Max number of images in one post (0 = images not supported).
        video:
          type: boolean
          description: Whether a (single) video is supported.
        mixedMedia:
          type: boolean
          description: Whether images and videos can be mixed in one post.
        maxMediaItems:
          type: integer
          description: Max number of media items when mixing is allowed.
        needsPublicMediaUrl:
          type: [boolean, string]
          enum: [false, images, all]
          description: Which media the platform downloads from the public media URL (so it must be reachable from the internet).
        usesTitle:
          type: boolean
          description: Whether the post title is used.
    OptionField:
      type: object
      required: [key, label, type]
      properties:
        key: { type: string }
        label: { type: string }
        type:
          type: string
          enum: [select, checkbox, text]
        choices:
          type: array
          items:
            type: object
            required: [value, label]
            properties:
              value: { type: string }
              label: { type: string }
        default:
          type: [string, boolean]
        help: { type: string }
        remember:
          type: boolean
          description: false = a UI must not remember this choice for the next post (e.g. TikTok privacy, per TikTok's rules).
    CredentialField:
      type: object
      required: [key, label, type]
      properties:
        key: { type: string }
        label: { type: string }
        type:
          type: string
          enum: [text, password, url]
        placeholder: { type: string }
        help: { type: string }
        required: { type: boolean }
    PlatformDescription:
      type: object
      required: [id, name, connector, capabilities, options]
      properties:
        id: { $ref: "#/components/schemas/PlatformId" }
        name: { type: string, examples: [Instagram] }
        connector: { $ref: "#/components/schemas/ConnectorId" }
        capabilities: { $ref: "#/components/schemas/Capabilities" }
        options:
          type: array
          description: Per-post options (send their values in `platformOptions` or `targets[].options`).
          items: { $ref: "#/components/schemas/OptionField" }
    ConnectorDescription:
      type: object
      required: [id, name, platforms, kind, configured, redirectUri, developerPortal, credentialFields]
      properties:
        id: { $ref: "#/components/schemas/ConnectorId" }
        name: { type: string, examples: [Facebook & Instagram] }
        platforms:
          type: array
          items: { $ref: "#/components/schemas/PlatformId" }
        kind:
          type: string
          enum: [oauth, credentials]
          description: "`oauth`: send the user to the platform; `credentials`: collect `credentialFields` in your UI."
        configured:
          type: boolean
          description: Whether the app keys for this platform are set.
        redirectUri:
          type: [string, "null"]
          format: uri
          description: Redirect URI to register in the platform's developer console (OAuth connectors).
        developerPortal:
          type: [string, "null"]
          format: uri
        credentialFields:
          type: [array, "null"]
          items: { $ref: "#/components/schemas/CredentialField" }
    Description:
      type: object
      required: [publicUrl, publicMediaReachable, ffmpeg, maxUploadMb, supportedMimeTypes, platforms, connectors]
      properties:
        publicUrl:
          type: string
          format: uri
          description: Public base URL of the handler (OAuth callbacks and media URLs live under it).
        publicMediaReachable:
          type: boolean
          description: False when publicUrl is a private address; Instagram photos and Threads media can't be fetched then.
        ffmpeg:
          type: boolean
          description: Whether ffmpeg/ffprobe are installed on the server.
        maxUploadMb: { type: integer }
        supportedMimeTypes:
          type: array
          items: { type: string }
          examples: [[image/jpeg, image/png, image/webp, image/gif, video/mp4, video/quicktime, video/webm, video/x-m4v]]
        platforms:
          type: array
          items: { $ref: "#/components/schemas/PlatformDescription" }
        connectors:
          type: array
          items: { $ref: "#/components/schemas/ConnectorDescription" }
...
    TargetIssue:
      type: object
      required: [accountId, accountName, platform, errors, length]
      properties:
        accountId: { type: string }
        accountName: { type: string }
        platform: { $ref: "#/components/schemas/PlatformId" }
        errors:
          type: array
          description: Empty when this account can take the post.
          items: { type: string }
        length:
          type: integer
          description: Caption length as this platform counts it.
...